Crate lettre

Lettre is a mailer written in Rust. It provides a simple email builder and several transports.

This mailer contains the available transports for your emails.

Re-exports

pub use smtp::client::net::ClientTlsParameters;

pub use smtp::ClientSecurity;

pub use smtp::SmtpClient;

pub use smtp::SmtpTransport;

Modules

error

smtp

The SMTP transport sends emails using the SMTP protocol.

stub

The stub transport only logs message envelope and drops the content. It can be useful for testing purposes.

Structs

EmailAddress

Email address

Envelope

Simple email envelope representation

SendableEmail

Sendable email structure

Enums

Message

Traits

Transport

Transport method for emails