Struct lettre::email::EmailBuilder

pub struct EmailBuilder {
    // some fields omitted
}

Builds an Email structure

Methods

impl EmailBuilder

fn new() -> EmailBuilder

Creates a new empty email

fn body(self, body: &str) -> EmailBuilder

Sets the email body

fn add_header<A: ToHeader>(self, header: A) -> EmailBuilder

Add a generic header

fn from<A: ToMailbox>(self, address: A) -> EmailBuilder

Adds a From header and store the sender address

fn to<A: ToMailbox>(self, address: A) -> EmailBuilder

Adds a To header and store the recipient address

fn cc<A: ToMailbox>(self, address: A) -> EmailBuilder

Adds a Cc header and store the recipient address

fn reply_to<A: ToMailbox>(self, address: A) -> EmailBuilder

Adds a Reply-To header

fn sender<A: ToMailbox>(self, address: A) -> EmailBuilder

Adds a Sender header

fn subject(self, subject: &str) -> EmailBuilder

Adds a Subject header

fn date(self, date: &Tm) -> EmailBuilder

Adds a Date header with the given date

fn build(self) -> Result<Email, &'static str>

Build the Email