lettre::message

Module header

Headers widely used in email messages

Structs

Bcc

Bcc header

Cc

Cc header

Comments

Comments of the message, defined in RFC5322

ContentDisposition

Content-Disposition of an attachment

ContentId

Content-Id header, defined in RFC2045

ContentLocation

Content-Location header, defined in RFC2110

ContentType

Content-Type of the body

ContentTypeErr

An error occurred while trying to ContentType::parse.

Date

Message Date header

From

From header

HeaderName

A valid header name

HeaderValue

A safe for use header value

Headers

A set of email headers

InReplyTo

In-Reply-To header. Contains one or more unique message identifiers, defined in RFC5322

InvalidHeaderName

A possible error when converting a HeaderName from another type.

Keywords

Keywords header. Should contain a comma-separated list of one or more words or quoted-strings, defined in RFC5322

MessageId

Message-Id header. Contains a unique message identifier, defined in RFC5322

MimeVersion

Message format version, defined in RFC2045

References

References header. Contains one or more unique message identifiers, defined in RFC5322

ReplyTo

Reply-To header

Sender

Sender header

Subject

Subject of the message, defined in RFC5322

To

To header

UserAgent

User-Agent header. Contains information about the client, defined in draft-melnikov-email-user-agent-00

Enums

ContentTransferEncoding

Content-Transfer-Encoding of the body

Constants

MIME_VERSION_1_0

MIME version 1.0

Traits

Header

Represents an email header

MailboxesHeader

Header which can contains multiple mailboxes