Struct lettre::message::Mailboxes

pub struct Mailboxes(_);

Represents a sequence of Mailbox instances.

This type contains a sequence of mailboxes (Some Name <user@domain.tld>, Another Name <other@domain.tld>, withoutname@domain.tld, …).

NOTE: Enable feature “serde” to be able serialize/deserialize it using serde.

Implementations

impl Mailboxes

pub fn new() -> Self

Creates a new list of Mailbox instances.

Examples

use lettre::message::Mailboxes;
let mailboxes = Mailboxes::new();

pub fn with(self, mbox: Mailbox) -> Self

Adds a new Mailbox to the list, in a builder style pattern.

Examples

use lettre::{
    message::{Mailbox, Mailboxes},
    Address,
};

let address = Address::new("example", "email.com")?;
let mut mailboxes = Mailboxes::new().with(Mailbox::new(None, address));

pub fn push(&mut self, mbox: Mailbox)

Adds a new Mailbox to the list, in a Vec::push style pattern.

Examples

use lettre::{
    message::{Mailbox, Mailboxes},
    Address,
};

let address = Address::new("example", "email.com")?;
let mut mailboxes = Mailboxes::new();
mailboxes.push(Mailbox::new(None, address));

pub fn into_single(self) -> Option<Mailbox>

Extracts the first Mailbox if it exists.

Examples

use lettre::{
    message::{Mailbox, Mailboxes},
    Address,
};

let empty = Mailboxes::new();
assert!(empty.into_single().is_none());

let mut mailboxes = Mailboxes::new();
let address = Address::new("example", "email.com")?;

mailboxes.push(Mailbox::new(None, address));
assert!(mailboxes.into_single().is_some());

pub fn iter(&self) -> Iter<'_, Mailbox>

Creates an iterator over the Mailbox instances that are currently stored.

Examples

use lettre::{
    message::{Mailbox, Mailboxes},
    Address,
};

let mut mailboxes = Mailboxes::new();

let address = Address::new("example", "email.com")?;
mailboxes.push(Mailbox::new(None, address));

let address = Address::new("example", "email.com")?;
mailboxes.push(Mailbox::new(None, address));

let mut iter = mailboxes.iter();

assert!(iter.next().is_some());
assert!(iter.next().is_some());

assert!(iter.next().is_none());