Struct lettre::message::MessageBuilder

pub struct MessageBuilder { /* private fields */ }

A builder for messages

Implementations

impl MessageBuilder

pub fn new() -> Self

Creates a new default message builder

pub fn from(self, mbox: Mailbox) -> Self

Set or add mailbox to From header

Defined in RFC5322.

Shortcut for self.mailbox(header::From(mbox)).

pub fn sender(self, mbox: Mailbox) -> Self

Set Sender header. Should be used when providing several From mailboxes.

Defined in RFC5322.

Shortcut for self.header(header::Sender(mbox)).

pub fn date(self, st: SystemTime) -> Self

Add Date header to message

Shortcut for self.header(header::Date::new(st)).

pub fn date_now(self) -> Self

Set Date header using current date/time

Shortcut for self.date(SystemTime::now()), it is automatically inserted if no date has been provided.

pub fn reply_to(self, mbox: Mailbox) -> Self

Set or add mailbox to ReplyTo header

Defined in RFC5322.

Shortcut for self.mailbox(header::ReplyTo(mbox)).

pub fn to(self, mbox: Mailbox) -> Self

Set or add mailbox to To header

Shortcut for self.mailbox(header::To(mbox)).

pub fn cc(self, mbox: Mailbox) -> Self

Set or add mailbox to Cc header

Shortcut for self.mailbox(header::Cc(mbox)).

pub fn bcc(self, mbox: Mailbox) -> Self

Set or add mailbox to Bcc header

Shortcut for self.mailbox(header::Bcc(mbox)).

pub fn in_reply_to(self, id: String) -> Self

Set or add message id to In-Reply-To header

pub fn references(self, id: String) -> Self

Set or add message id to References header

pub fn subject<S: Into<String>>(self, subject: S) -> Self

Set Subject header to message

Shortcut for self.header(header::Subject(subject.into())).

pub fn message_id(self, id: Option<String>) -> Self

Set Message-ID header

Should generally be inserted by the mail relay.

If None is provided, an id will be generated in the <UUID@HOSTNAME>.

pub fn user_agent(self, id: String) -> Self

Set User-Agent header

pub fn header<H: Header>(self, header: H) -> Self

Set custom header to message

pub fn mailbox<H: Header + MailboxesHeader>(self, header: H) -> Self

Add mailbox to header

pub fn envelope(self, envelope: Envelope) -> Self

Force specific envelope (by default it is derived from headers)

pub fn keep_bcc(self) -> Self

Keep the Bcc header

By default the Bcc header is removed from the email after using it to generate the message envelope. In some cases though, like when saving the email as an .eml, or sending through some transports (like the Gmail API) that don’t take a separate envelope value, it becomes necessary to keep the Bcc header.

Calling this method overrides the default behaviour.

pub fn body<T: IntoBody>(self, body: T) -> Result<Message, EmailError>

Create Message using a Vec<u8>, String, or Body body

Automatically gets encoded with 7bit, quoted-printable or base64 Content-Transfer-Encoding, based on the most efficient and valid encoding for body.

pub fn multipart(self, part: MultiPart) -> Result<Message, EmailError>

Create message using mime body (MultiPart)

pub fn singlepart(self, part: SinglePart) -> Result<Message, EmailError>

Create message using mime body (SinglePart)