Module lettre::smtp

The SMTP transport sends emails using the SMTP protocol.

This SMTP client follows RFC 5321, and is designed to efficiently send emails from an application to a relay email server, as it relies as much as possible on the relay server for sanity and RFC compliance checks.

It implements the following extensions:

8BITMIME (RFC 6152)

AUTH (RFC 4954) with PLAIN, LOGIN and XOAUTH2 mechanisms

STARTTLS (RFC 2487)

SMTPUTF8 (RFC 6531)

Modules

authentication

Provides limited SASL authentication mechanisms

client

SMTP client

commands

SMTP commands

error

Error and result type for SMTP clients

extension

ESMTP features

response

SMTP response, containing a mandatory return code and an optional text message

util

Utils for string manipulation

Structs

SmtpClient

Contains client configuration

SmtpTransport

Structure that implements the high level SMTP client

Enums

ClientSecurity

How to apply TLS to a client connection

ConnectionReuseParameters

Configures connection reuse behavior

Constants

SMTP_PORT

Default smtp port

SUBMISSIONS_PORT

Default submission over TLS port

SUBMISSION_PORT

Default submission port