Struct lettre::smtp::client::Client

pub struct Client<S: Write + Read = NetworkStream> { /* fields omitted */ }

Structure that implements the SMTP client

Methods

impl<S: Write + Read> Client<S>

pub fn new() -> Client<S>

Creates a new SMTP client

It does not connects to the server, but only creates the Client

impl<S: Connector + Write + Read + Timeout + Debug> Client<S>

pub fn close(&mut self)

Closes the SMTP transaction if possible

pub fn set_stream(&mut self, stream: S)

Sets the underlying stream

pub fn upgrade_tls_stream(
&mut self,
tls_parameters: &ClientTlsParameters
) -> Result<()>

Upgrades the underlying connection to SSL/TLS

pub fn is_encrypted(&self) -> bool

Tells if the underlying stream is currently encrypted

pub fn set_timeout(&mut self, duration: Option<Duration>) -> Result<()>

Set timeout

pub fn connect<A: ToSocketAddrs>(
&mut self,
addr: &A,
tls_parameters: Option<&ClientTlsParameters>
) -> SmtpResult

Connects to the configured server

pub fn is_connected(&mut self) -> bool

Checks if the server is connected using the NOOP SMTP command

pub fn auth(
&mut self,
mechanism: Mechanism,
credentials: &Credentials
) -> SmtpResult

Sends an AUTH command with the given mechanism, and handles challenge if needed

pub fn message<T: Read>(&mut self, message: Box<T>) -> SmtpResult

Sends the message content

pub fn smtp_command<C: Display>(&mut self, command: C) -> SmtpResult

Sends an SMTP command