Struct lettre::smtp::response::Response

pub struct Response {
    pub code: Code,
    pub message: Vec<String>,
}

Contains an SMTP reply, with separated code and message

The text message is optional, only the code is mandatory

Fields
code: Code

Response code

message: Vec<String>

Server response string (optional) Handle multiline responses

Methods

impl Response

pub fn new(code: Code, message: Vec<String>) -> Response

Creates a new Response

pub fn is_positive(&self) -> bool

Tells if the response is positive

pub fn has_code(&self, code: u16) -> bool

Tests code equality

pub fn first_word(&self) -> Option<&str>

Returns only the first word of the message if possible

pub fn first_line(&self) -> Option<&str>

Returns only the line of the message if possible