Struct lettre::smtp::SmtpClient

pub struct SmtpClient { /* fields omitted */ }

Contains client configuration

Implementations

impl SmtpClient

Builder for the SMTP SmtpTransport

pub fn new<A: ToSocketAddrs>(
addr: A,
security: ClientSecurity
) -> Result<SmtpClient, Error>

Creates a new SMTP client

Defaults are:

No connection reuse

No authentication

No SMTPUTF8 support

A 60 seconds timeout for smtp commands

Consider using SmtpClient::new_simple instead, if possible.

pub fn new_simple(domain: &str) -> Result<SmtpClient, Error>

Simple and secure transport, should be used when possible. Creates an encrypted transport over submissions port, using the provided domain to validate TLS certificates.

pub fn new_unencrypted_localhost() -> Result<SmtpClient, Error>

Creates a new local SMTP client to port 25

pub fn smtp_utf8(self, enabled: bool) -> SmtpClient

Enable SMTPUTF8 if the server supports it

pub fn hello_name(self, name: ClientId) -> SmtpClient

Set the name used during EHLO

pub fn connection_reuse(
self,
parameters: ConnectionReuseParameters
) -> SmtpClient

Enable connection reuse

pub fn credentials<S: Into<Credentials>>(self, credentials: S) -> SmtpClient

Set the client credentials

pub fn authentication_mechanism(self, mechanism: Mechanism) -> SmtpClient

Set the authentication mechanism to use

pub fn timeout(self, timeout: Option<Duration>) -> SmtpClient

Set the timeout duration

pub fn transport(self) -> SmtpTransport

Build the SMTP client

It does not connect to the server, but only creates the SmtpTransport