Struct lettre::smtp::SmtpTransport

pub struct SmtpTransport { /* fields omitted */ }

Structure that implements the high level SMTP client

Methods

impl<'a> SmtpTransport

pub fn simple_builder(domain: String) -> Result<SmtpTransportBuilder, Error>

Simple and secure transport, should be used when possible. Creates an encrypted transport over submission port, using the provided domain to validate TLS certificates.

pub fn builder<A: ToSocketAddrs>(
addr: A,
security: ClientSecurity
) -> Result<SmtpTransportBuilder, Error>

Creates a new configurable builder

pub fn builder_unencrypted_localhost() -> Result<SmtpTransportBuilder, Error>

Creates a new local SMTP client to port 25

pub fn new(builder: SmtpTransportBuilder) -> SmtpTransport

Creates a new SMTP client

It does not connect to the server, but only creates the SmtpTransport

pub fn get_ehlo(&mut self) -> SmtpResult

Gets the EHLO response and updates server information

pub fn close(&mut self)

Closes the inner connection

pub fn reset(&mut self)

Reset the client state