Struct lettre::SmtpTransport

pub struct SmtpTransport { /* private fields */ }

Sends emails using the SMTP protocol

Implementations

impl SmtpTransport

pub fn relay(relay: &str) -> Result<SmtpTransportBuilder, Error>

Simple and secure transport, using TLS connections to communicate with the SMTP server

The right option for most SMTP servers.

Creates an encrypted transport over submissions port, using the provided domain to validate TLS certificates.

pub fn starttls_relay(relay: &str) -> Result<SmtpTransportBuilder, Error>

Simple an secure transport, using STARTTLS to obtain encrypted connections

Alternative to SmtpTransport::relay, for SMTP servers that don’t take SMTPS connections.

Creates an encrypted transport over submissions port, by first connecting using an unencrypted connection and then upgrading it with STARTTLS. The provided domain is used to validate TLS certificates.

An error is returned if the connection can’t be upgraded. No credentials or emails will be sent to the server, protecting from downgrade attacks.

pub fn unencrypted_localhost() -> SmtpTransport

Creates a new local SMTP client to port 25

Shortcut for local unencrypted relay (typical local email daemon that will handle relaying)

pub fn builder_dangerous<T: Into<String>>(server: T) -> SmtpTransportBuilder

Creates a new SMTP client

Defaults are:

No authentication

No TLS

A 60 seconds timeout for smtp commands

Port 25

Consider using SmtpTransport::relay or SmtpTransport::starttls_relay instead, if possible.

pub fn test_connection(&self) -> Result<bool, Error>

Tests the SMTP connection

test_connection() tests the connection by using the SMTP NOOP command. The connection is closed afterwards if a connection pool is not used.