Enum lettre::transport::smtp::authentication::Mechanism

pub enum Mechanism {
    Plain,
    CramMd5,
}

Represents authentication mechanisms

Variants

Plain

PLAIN authentication mechanism RFC 4616: https://tools.ietf.org/html/rfc4616

CramMd5

CRAM-MD5 authentication mechanism RFC 2195: https://tools.ietf.org/html/rfc2195

Methods

impl Mechanism

fn supports_initial_response(&self) -> bool

Does the mechanism supports initial response

fn response(&self, username: &str, password: &str, challenge: Option<&str>) -> Result<String, Error>

Returns the string to send to the server, using the provided username, password and challenge in some cases