Struct lettre::transport::smtp::client::Client

pub struct Client<S: Write + Read = NetworkStream> {
    // some fields omitted
}

Structure that implements the SMTP client

Methods

impl<S: Write + Read = NetworkStream> Client<S>

fn new() -> Client<S>

Creates a new SMTP client

It does not connects to the server, but only creates the Client

impl<S: Connector + Write + Read + Debug + Clone = NetworkStream> Client<S>

fn close(&mut self)

Closes the SMTP transaction if possible

fn set_stream(&mut self, stream: S)

Sets the underlying stream

fn upgrade_tls_stream(&mut self, ssl_context: &SslContext) -> Result<()>

Upgrades the underlying connection to SSL/TLS

fn connect<A: ToSocketAddrs>(&mut self, addr: &A) -> EmailResult

Connects to the configured server

fn is_connected(&mut self) -> bool

Checks if the server is connected using the NOOP SMTP command

fn command(&mut self, command: &str) -> EmailResult

Sends an SMTP command

fn ehlo(&mut self, hostname: &str) -> EmailResult

Sends a EHLO command

fn mail(&mut self, address: &str, options: Option<&str>) -> EmailResult

Sends a MAIL command

fn rcpt(&mut self, address: &str) -> EmailResult

Sends a RCPT command

fn data(&mut self) -> EmailResult

Sends a DATA command

fn quit(&mut self) -> EmailResult

Sends a QUIT command

fn noop(&mut self) -> EmailResult

Sends a NOOP command

fn help(&mut self, argument: Option<&str>) -> EmailResult

Sends a HELP command

fn vrfy(&mut self, address: &str) -> EmailResult

Sends a VRFY command

fn expn(&mut self, address: &str) -> EmailResult

Sends a EXPN command

fn rset(&mut self) -> EmailResult

Sends a RSET command

fn auth(&mut self, mecanism: Mecanism, username: &str, password: &str) -> EmailResult

Sends an AUTH command with the given mecanism

fn starttls(&mut self) -> EmailResult

Sends a STARTTLS command

fn message(&mut self, message_content: &str) -> EmailResult

Sends the message content