Struct lettre::transport::smtp::response::Response

pub struct Response {
    // some fields omitted
}

Contains an SMTP reply, with separed code and message

The text message is optional, only the code is mandatory

Methods

impl Response

fn new(code: Code, message: Vec<String>) -> Response

Creates a new Response

fn is_positive(&self) -> bool

Tells if the response is positive

fn message(&self) -> Vec<String>

Returns the message

fn severity(&self) -> Severity

Returns the severity (i.e. 1st digit)

fn category(&self) -> Category

Returns the category (i.e. 2nd digit)

fn detail(&self) -> u8

Returns the detail (i.e. 3rd digit)

fn has_code(&self, code: u16) -> bool

Tests code equality

fn first_word(&self) -> Option<String>

Returns only the first word of the message if possible