Struct lettre::transport::smtp::SmtpTransportBuilder

pub struct SmtpTransportBuilder {
    // some fields omitted
}

Contains client configuration

Methods

impl SmtpTransportBuilder

Builder for the SMTP SmtpTransport

fn new<A: ToSocketAddrs>(addr: A) -> Result<SmtpTransportBuilder, Error>

Creates a new local SMTP client

fn localhost() -> Result<SmtpTransportBuilder, Error>

Creates a new local SMTP client to port 25

fn ssl_context(self, ssl_context: SslContext) -> SmtpTransportBuilder

Use STARTTLS with a specific context

fn security_level(self, level: SecurityLevel) -> SmtpTransportBuilder

Require SSL/TLS using STARTTLS

fn smtp_utf8(self, enabled: bool) -> SmtpTransportBuilder

Require SSL/TLS using STARTTLS

fn hello_name(self, name: &str) -> SmtpTransportBuilder

Set the name used during HELO or EHLO

fn connection_reuse(self, enable: bool) -> SmtpTransportBuilder

Enable connection reuse

fn connection_reuse_count_limit(self, limit: u16) -> SmtpTransportBuilder

Set the maximum number of emails sent using one connection

fn credentials(self, username: &str, password: &str) -> SmtpTransportBuilder

Set the client credentials

fn authentication_mecanisms(self, mecanisms: Vec<Mecanism>) -> SmtpTransportBuilder

Set the authentication mecanisms

fn build(self) -> SmtpTransport

Build the SMTP client

It does not connects to the server, but only creates the SmtpTransport