Crate lettre

Lettre is a mailer written in Rust. It provides a simple email builder and several transports.

This mailer contains the available transports for your emails. To be sendable, the emails have to implement SendableEmail.

Re-exports

pub use file::FileEmailTransport;

pub use sendmail::SendmailTransport;

pub use smtp::SmtpTransport;

pub use smtp::ClientSecurity;

pub use smtp::client::net::ClientTlsParameters;

Modules

file
The file transport writes the emails to the given directory. The name of the file will be message_id.txt. It can be useful for testing purposes, or if you want to keep track of sent messages.

sendmail
The sendmail transport sends the email using the local sendmail command.

smtp
The SMTP transport sends emails using the SMTP protocol.

stub
The stub transport only logs message envelope and drops the content. It can be useful for testing purposes.

Structs

EmailAddress
Email address

SimpleSendableEmail
Minimal email structure

Traits

EmailTransport
Transport method for emails

SendableEmail
Email sendable by an SMTP client