lettre::address
Struct Address

pub struct Address { /* private fields */ }

Represents an email address with a user and a domain name.

This type contains email in canonical form (user@domain.tld).

NOTE: Enable feature “serde” to be able to serialize/deserialize it using serde.

Examples

You can create an Address from a user and a domain:

use lettre::Address;

let address = Address::new("user", "email.com")?;
assert_eq!(address.user(), "user");
assert_eq!(address.domain(), "email.com");

You can also create an Address from a string literal by parsing it:

use lettre::Address;

let address = "user@email.com".parse::<Address>()?;
assert_eq!(address.user(), "user");
assert_eq!(address.domain(), "email.com");

Implementations

impl Address

pub fn new<U: AsRef<str>, D: AsRef<str>>( user: U, domain: D, ) -> Result<Self, AddressError>

Creates a new email address from a user and domain.

Examples

use lettre::Address;

let address = Address::new("user", "email.com")?;
let expected = "user@email.com".parse::<Address>()?;
assert_eq!(expected, address);

pub fn user(&self) -> &str

Gets the user portion of the Address.

Examples

use lettre::Address;

let address = Address::new("user", "email.com")?;
assert_eq!(address.user(), "user");

pub fn domain(&self) -> &str

Gets the domain portion of the Address.

Examples

use lettre::Address;

let address = Address::new("user", "email.com")?;
assert_eq!(address.domain(), "email.com");