Struct lettre::email::EmailBuilder

pub struct EmailBuilder {
    // some fields omitted
}

Builds an Email structure

Methods

impl EmailBuilder

fn new() -> EmailBuilder

Creates a new empty email

fn body<S: Into<String>>(self, body: S) -> EmailBuilder

Sets the email body

fn set_body<S: Into<String>>(&mut self, body: S)

Sets the email body

fn header<A: IntoHeader>(self, header: A) -> EmailBuilder

Add a generic header

fn add_header<A: IntoHeader>(&mut self, header: A)

Add a generic header

fn from<A: IntoMailbox>(self, address: A) -> EmailBuilder

Adds a From header and stores the sender address

fn add_from<A: IntoMailbox>(&mut self, address: A)

Adds a From header and stores the sender address

fn to<A: IntoMailbox>(self, address: A) -> EmailBuilder

Adds a To header and stores the recipient address

fn add_to<A: IntoMailbox>(&mut self, address: A)

Adds a To header and stores the recipient address

fn cc<A: IntoMailbox>(self, address: A) -> EmailBuilder

Adds a Cc header and stores the recipient address

fn add_cc<A: IntoMailbox>(&mut self, address: A)

Adds a Cc header and stores the recipient address

fn bcc<A: IntoMailbox>(self, address: A) -> EmailBuilder

Adds a Bcc header and stores the recipient address

fn add_bcc<A: IntoMailbox>(&mut self, address: A)

Adds a Bcc header and stores the recipient address

fn reply_to<A: IntoMailbox>(self, address: A) -> EmailBuilder

Adds a Reply-To header

fn add_reply_to<A: IntoMailbox>(&mut self, address: A)

Adds a Reply-To header

fn sender<A: IntoMailbox>(self, address: A) -> EmailBuilder

Adds a Sender header

fn set_sender<A: IntoMailbox>(&mut self, address: A)

Adds a Sender header

fn subject<S: Into<String>>(self, subject: S) -> EmailBuilder

Adds a Subject header

fn set_subject<S: Into<String>>(&mut self, subject: S)

Adds a Subject header

fn date(self, date: &Tm) -> EmailBuilder

Adds a Date header with the given date

fn set_date(&mut self, date: &Tm)

Adds a Date header with the given date

fn message_type(self, message_type: MimeMultipartType) -> EmailBuilder

Set the message type

fn set_message_type(&mut self, message_type: MimeMultipartType)

Set the message type

fn child(self, child: MimeMessage) -> EmailBuilder

Adds a child

fn add_child(&mut self, child: MimeMessage)

Adds a child

fn text<S: Into<String>>(self, body: S) -> EmailBuilder

Sets the email body to plain text content

fn set_text<S: Into<String>>(&mut self, body: S)

Sets the email body to plain text content

fn html<S: Into<String>>(self, body: S) -> EmailBuilder

Sets the email body to HTML content

fn set_html<S: Into<String>>(&mut self, body: S)

Sets the email body to HTML content

fn alternative<S: Into<String>, T: Into<String>>(self, body_html: S, body_text: T) -> EmailBuilder

Sets the email content

fn set_alternative<S: Into<String>, T: Into<String>>(&mut self, body_html: S, body_text: T)

Sets the email content

fn envelope(self, envelope: Envelope) -> EmailBuilder

Sets the envelope for manual destination control If this function is not called, the envelope will be calculated from the "to" and "cc" addresses you set.

fn set_envelope(&mut self, envelope: Envelope)

Sets the envelope for manual destination control If this function is not called, the envelope will be calculated from the "to" and "cc" addresses you set.

fn build(self) -> Result<Email, Error>

Builds the Email