Struct lettre::message::Attachment

pub struct Attachment { /* private fields */ }

SinglePart builder for attachments

Allows building attachment parts easily.

Implementations

impl Attachment

pub fn new(filename: String) -> Self

Create a new attachment

This attachment will be displayed as a normal attachment, with the chosen filename appearing as the file name.

use std::fs;

use lettre::message::{header::ContentType, Attachment};

let filename = String::from("invoice.pdf");
let filebody = fs::read("invoice.pdf")?;
let content_type = ContentType::parse("application/pdf").unwrap();
let attachment = Attachment::new(filename).body(filebody, content_type);

// The document `attachment` will show up as a normal attachment.

pub fn new_inline(content_id: String) -> Self

Create a new inline attachment

This attachment should be displayed inline into the message body:

<img src="cid:123">

use std::fs;

use lettre::message::{header::ContentType, Attachment};

let content_id = String::from("123");
let filebody = fs::read("image.jpg")?;
let content_type = ContentType::parse("image/jpeg").unwrap();
let attachment = Attachment::new_inline(content_id).body(filebody, content_type);

// The image `attachment` will display inline into the email.

pub fn body<T: IntoBody>(
self,
content: T,
content_type: ContentType
) -> SinglePart

Build the attachment into a SinglePart which can then be used to build the rest of the email

Look at the Complex MIME body example to see how SinglePart can be put into the email.