Enum lettre::transport::smtp::authentication::Mechanism

pub enum Mechanism {
    Plain,
    Login,
    Xoauth2,
}

Represents authentication mechanisms

Variants

Plain

PLAIN authentication mechanism, defined in RFC 4616

Login

LOGIN authentication mechanism Obsolete but needed for some providers (like Office 365)

Defined in draft-murchison-sasl-login-00.

Xoauth2

Non-standard XOAUTH2 mechanism, defined in xoauth2-protocol

Implementations

impl Mechanism

pub fn supports_initial_response(self) -> bool

Does the mechanism support initial response?

pub fn response( self, credentials: &Credentials, challenge: Option<&str>, ) -> Result<String, Error>

Returns the string to send to the server, using the provided username, password and challenge in some cases