Struct lettre::transport::smtp::client::SmtpConnection

pub struct SmtpConnection { /* private fields */ }

Structure that implements the SMTP client

Implementations

impl SmtpConnection

pub fn server_info(&self) -> &ServerInfo

Get information about the server

pub fn connect<A: ToSocketAddrs>( server: A, timeout: Option<Duration>, hello_name: &ClientId, tls_parameters: Option<&TlsParameters>, local_address: Option<IpAddr> ) -> Result<SmtpConnection, Error>

Connects to the configured server

Sends EHLO and parses server information

pub fn send( &mut self, envelope: &Envelope, email: &[u8] ) -> Result<Response, Error>

pub fn has_broken(&self) -> bool

pub fn can_starttls(&self) -> bool

pub fn starttls( &mut self, tls_parameters: &TlsParameters, hello_name: &ClientId ) -> Result<(), Error>

pub fn quit(&mut self) -> Result<Response, Error>

pub fn abort(&mut self)

pub fn set_stream(&mut self, stream: NetworkStream)

Sets the underlying stream

pub fn is_encrypted(&self) -> bool

Tells if the underlying stream is currently encrypted

pub fn set_timeout(&mut self, duration: Option<Duration>) -> Result<()>

Set timeout

pub fn test_connected(&mut self) -> bool

Checks if the server is connected using the NOOP SMTP command

pub fn auth( &mut self, mechanisms: &[Mechanism], credentials: &Credentials ) -> Result<Response, Error>

Sends an AUTH command with the given mechanism, and handles challenge if needed

pub fn message(&mut self, message: &[u8]) -> Result<Response, Error>

Sends the message content

pub fn command<C: Display>(&mut self, command: C) -> Result<Response, Error>

Sends an SMTP command

pub fn read_response(&mut self) -> Result<Response, Error>

Gets the SMTP response

pub fn peer_certificate(&self) -> Result<Vec<u8>, Error>

The X509 certificate of the server (DER encoded)