Struct lettre::transport::smtp::SmtpTransportBuilder

pub struct SmtpTransportBuilder { /* fields omitted */ }

Contains client configuration

Methods

impl SmtpTransportBuilder

Builder for the SMTP SmtpTransport

fn new<A: ToSocketAddrs>(addr: A) -> Result<SmtpTransportBuilder, Error>

Creates a new local SMTP client

fn localhost() -> Result<SmtpTransportBuilder, Error>

Creates a new local SMTP client to port 25

fn ssl_context(self, ssl_context: SslContext) -> SmtpTransportBuilder

Use STARTTLS with a specific context

fn security_level(self, level: SecurityLevel) -> SmtpTransportBuilder

Set the security level for SSL/TLS

fn encrypt(self) -> SmtpTransportBuilder

Require SSL/TLS using STARTTLS

Incompatible with ssl_wrapper()`

fn ssl_wrapper(self) -> SmtpTransportBuilder

Require SSL/TLS using SMTPS

Incompatible with encrypt()

fn smtp_utf8(self, enabled: bool) -> SmtpTransportBuilder

Enable SMTPUTF8 if the server supports it

fn hello_name<S: Into<String>>(self, name: S) -> SmtpTransportBuilder

Set the name used during HELO or EHLO

fn connection_reuse(self, enable: bool) -> SmtpTransportBuilder

Enable connection reuse

fn connection_reuse_count_limit(self, limit: u16) -> SmtpTransportBuilder

Set the maximum number of emails sent using one connection

fn credentials<S: Into<String>>(self,
username: S,
password: S)
-> SmtpTransportBuilder

Set the client credentials

fn authentication_mechanism(self, mechanism: Mechanism) -> SmtpTransportBuilder

Set the authentication mechanisms

fn timeout(self, timeout: Option<Duration>) -> SmtpTransportBuilder

Set the timeout duration

fn build(self) -> SmtpTransport

Build the SMTP client

It does not connect to the server, but only creates the SmtpTransport