Struct lettre::transport::smtp::SmtpTransportBuilder

pub struct SmtpTransportBuilder { /* private fields */ }

Contains client configuration. Instances of this struct can be created using functions of SmtpTransport.

Implementations

impl SmtpTransportBuilder

Builder for the SMTP SmtpTransport

pub fn hello_name(self, name: ClientId) -> Self

Set the name used during EHLO

pub fn credentials(self, credentials: Credentials) -> Self

Set the authentication mechanism to use

pub fn authentication(self, mechanisms: Vec<Mechanism>) -> Self

Set the authentication mechanism to use

pub fn timeout(self, timeout: Option<Duration>) -> Self

Set the timeout duration

pub fn port(self, port: u16) -> Self

Set the port to use

pub fn tls(self, tls: Tls) -> Self

Set the TLS settings to use

pub fn pool_config(self, pool_config: PoolConfig) -> Self

Use a custom configuration for the connection pool

Defaults can be found at PoolConfig

pub fn build(self) -> SmtpTransport

Build the transport

If the pool feature is enabled, an Arc wrapped pool is created. Defaults can be found at PoolConfig